v0.3

+++ server/app/Http/Controllers/CategoriasController.php
@@ -3,42 +3,69 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Categoria;
 use App\Http\Requests\CreateCategoriaRequest;
 use App\Http\Requests\EditCategoriaRequest;
-use App\Categoria;
 
 class CategoriasController extends Controller
 {
-    public function index(){
-        return response()->json(Categoria::all()->produtos);
-    }
-
-    public function store(CreateCategoriaRequest $request){
-        $categoria = Categoria::create($request->all());
-        $categoria->save();
-        return response()->json([
-            'mensagem' => 'Categoria cadastrada com sucesso!'
-        ], 201);
-    }
-
-    public function show($id){
-    }
-
-    public function update(EditCategoriaRequest $request, $id){
-        $categoria = Categoria::find($id);
-        $categoria->update([
-            'name' => $request->name
-        ]);
-        return response()->json([
-            'mensagem' => 'Categoria atualizada com sucesso'
-        ], 200);
+    public function __construct(){
     }
     
-    public function destroy($id){
+    public function index()
+    {
+        return view('categorias.index')->with('categorias',Categoria::all());
+    }
+
+    public function create()
+    {
+        return view('categorias.create');
+    }
+
+    public function store(CreateCategoriaRequest $request)
+    {
+        Categoria::create($request->all());
+        session()->flash('success', 'Categoria criada com sucesso!');
+        return redirect(route('categorias.index'));
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        //
+    }
+
+    public function edit(Categoria $categoria)
+    {
+        return view('categorias.edit')->with('categoria', $categoria);
+    }
+
+    public function update(EditCategoriaRequest $request, Categoria $categoria)
+    {
+        $categoria->update([
+            'nome' => $request->name
+        ]);
+
+        session()->flash('success', 'Categoria alterada com sucesso!');
+        return redirect(route('categorias.index'));
+    }
+
+    public function destroy($id)
+    {
         $categoria = Categoria::find($id);
+
+        $produto_count = $categoria->produtos()->count();
+        if($categoria->produtos()->count() > 0){
+            session()->flash('error', 'Não pode ser apagada pois temos ('.$produto_count.') produto(s) com essa categoria');
+            return redirect()->back();
+        }
         $categoria->delete();
-        return response()->json([
-            'mensagem' => 'Categoria excluída com sucesso'
-        ], 200);
+        session()->flash('success', 'Categoria movida para lixeira com sucesso!');
+        return redirect()->back();
     }
 }
